requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.6